pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowledge-engine"
version = "0.1.0"
description = "Continuous knowledge accumulation with structured evaluation loops"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Vache Sarkissian" }]
requires-python = ">=3.11"
keywords = ["knowledge", "research", "ai", "rag", "evaluation", "learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.1",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "sqlite-utils>=3.36",
    "rich>=13.7",
    "PyYAML>=6.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.26",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.23"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.9",
    "types-PyYAML",
]
all = ["knowledge-engine[anthropic,dev]"]

[project.scripts]
ke = "knowledge_engine.cli:main"

[project.urls]
Homepage = "https://github.com/vachsark/knowledge-engine"
Blog = "https://vachsark.com/blog"
Issues = "https://github.com/vachsark/knowledge-engine/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_engine"]

# --- ruff ---
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

# --- mypy ---
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

# --- pytest ---
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
